Adds Controller layer and fixes Repository pattern implementation

+++ src/main/java/com/gildedrose/items/Ageable.java
@@ -6,5 +6,5 @@ public interface Ageable {
 
     String getName();
 
-    int getPrice();
+    float getPrice();
 }
